requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.2.0
backoff==2.2.1
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
coinbase-advanced-py==1.4.2
cryptography==42.0.5
discord.py @ git+https://github.com/Rapptz/discord.py.git@2e2f20c6d5a7d938c1f9546dcf9a187694272e9a
frozenlist==1.4.1
idna==3.7
multidict==6.0.5
pycparser==2.22
PyJWT==2.8.0
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.3
typing_extensions==4.12.2
urllib3==2.2.2
websockets==12.0
yarl==1.9.4
